temporal-example/pyproject.toml
[project]
name = "temporal-example"
version = "0.1.0"
description = "Temporal Durable Execution Evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.9.0",
    "openai>=1.60.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
